pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "instagram-scraper"
version = "0.1.0"
description = "Scraper de perfiles de Instagram con Instaloader"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "instaloader>=4.10",
  "python-dotenv>=1.0",
  "playwright>=1.47",
  "cryptography>=42.0"
]
authors = [
  { name = "Proyecto" }
]

[project.scripts]
instagram-scraper = "instagram_scraper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/instagram_scraper"]